LayoutTests/fast/css-grid-layout/grid-item-column-row-get-set.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/grid-item-column-row-parsing-utils.js"></script>
<script>
if (window.testRunner)
    testRunner.overridePreference("WebKitCSSGridLayoutEnabled", 1);
</script>
<style>
.gridItemWithPositiveInteger {
    -webkit-grid-column: 10;
    -webkit-grid-row: 15;
}
.gridItemWithNegativeInteger {
    -webkit-grid-column: -10;
    -webkit-grid-row: -15;
}
.gridItemWithAuto {
    -webkit-grid-column: auto;
    -webkit-grid-row: auto;
}
.gridItemWith2Integer {
    -webkit-grid-column: 10 / 15;
    -webkit-grid-row: 5 / 5;
}
.gridItemWithNegativePositiveInteger {
    -webkit-grid-column: 10 / -10;
    -webkit-grid-row: -8 / 5;
}
.gridItemWithBeforeSpan {
    -webkit-grid-column: span 2 / 4;
    -webkit-grid-row: 3 / span 5;
}
.gridItemWithAfterSpan {
    -webkit-grid-column: 2 span / 4;
    -webkit-grid-row: 3 / 5 span;
}
.gridItemWith2OnlySpan {
    -webkit-grid-column: span / span;
    -webkit-grid-row: span / span;
}
.gridItemWith2Auto {
    -webkit-grid-column: auto / auto;
    -webkit-grid-row: auto / auto;
}
.gridItemWithBothLongHand {
    -webkit-grid-column-end: 11;
    -webkit-grid-row-start: 4;
}
.gridItemWithNoSpace {
    -webkit-grid-column: auto/1;
    -webkit-grid-row: 5/auto;
}
</style>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<!-- The first has no properties set on it. -->
<div id="gridElement"></div>
<div class="gridItemWithPositiveInteger" id="gridItemWithPositiveInteger"></div>
<div class="gridItemWithNegativeInteger" id="gridItemWithNegativeInteger"></div>
<div class="gridItemWithAuto" id="gridItemWithAutoElement"></div>
<div class="gridItemWith2Integer" id="gridItemWith2IntegerElement"></div>
<div class="gridItemWithNegativePositiveInteger" id="gridItemWithNegativePositiveIntegerElement"></div>
<div class="gridItemWith2Auto" id="gridItemWith2AutoElement"></div>
<div class="gridItemWithBeforeSpan" id="gridItemWithBeforeSpanElement"></div>
<div class="gridItemWithAfterSpan" id="gridItemWithAfterSpanElement"></div>
<div class="gridItemWith2OnlySpan" id="gridItemWith2OnlySpanElement"></div>
<div class="gridItemWithNegativePositiveInteger gridItemWithBothLongHand" id="gridItemWithBothShortLongHandElement"></div>
<div class="gridItemWithNoSpace" id="gridItemWithNoSpaceElement"></div>
<script>
    description('Test that setting and getting grid-column and grid-row works as expected');

    debug("Test getting -webkit-grid-column and -webkit-grid-row set through CSS");
    testColumnRowCSSParsing("gridElement", "auto / auto", "auto / auto");
    testColumnRowCSSParsing("gridItemWithPositiveInteger", "10 / auto", "15 / auto");
    testColumnRowCSSParsing("gridItemWithNegativeInteger", "-10 / auto", "-15 / auto");
    testColumnRowCSSParsing("gridItemWithAutoElement", "auto / auto", "auto / auto");
    testColumnRowCSSParsing("gridItemWith2IntegerElement", "10 / 15", "5 / 5");
    testColumnRowCSSParsing("gridItemWithNegativePositiveIntegerElement", "10 / -10", "-8 / 5");
    testColumnRowCSSParsing("gridItemWithBeforeSpanElement", "span 2 / 4", "3 / span 5");
    testColumnRowCSSParsing("gridItemWithAfterSpanElement", "span 2 / 4", "3 / span 5");
    testColumnRowCSSParsing("gridItemWith2OnlySpanElement", "auto / auto", "auto / auto");
    testColumnRowCSSParsing("gridItemWith2AutoElement", "auto / auto" , "auto / auto");
    testColumnRowCSSParsing("gridItemWithBothShortLongHandElement", "10 / 11", "4 / 5");
    testColumnRowCSSParsing("gridItemWithNoSpaceElement", "auto / 1", "5 / auto");

    debug("");
    debug("Test the initial value");
    var element = document.createElement("div");
    document.body.appendChild(element);
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column')", "'auto / auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column-start')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column-end')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row')", "'auto / auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row-start')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row-end')", "'auto'");

    debug("");
    debug("Test getting and setting grid-column and grid-row through JS");
    testColumnRowJSParsing("18", "66", "18 / auto", "66 / auto");
    testColumnRowJSParsing("-55", "-40", "-55 / auto", "-40 / auto");
    testColumnRowJSParsing("auto", "auto", "auto / auto", "auto / auto");
    testColumnRowJSParsing("10 / 55", "1 / 10");
    testColumnRowJSParsing("span 5 / 5", "4 / span 4");
    testColumnRowJSParsing("-5 / 5", "4 / -4");
    testColumnRowJSParsing("4 / auto", "5 / auto");
    testColumnRowJSParsing("auto / 5", "auto / 8");
    testColumnRowJSParsing("span / 3", "5 / span", "span 1 / 3", "5 / span 1");

    debug("");
    debug("Test setting grid-column and grid-row back to 'auto' through JS");
    element.style.webkitGridColumn = "18 / 19";
    element.style.webkitGridRow = "66 / 68";
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column')", "'18 / 19'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column-start')", "'18'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column-end')", "'19'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row')", "'66 / 68'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row-start')", "'66'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row-end')", "'68'");

    element.style.webkitGridColumn = "auto";
    element.style.webkitGridRow = "auto";
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column')", "'auto / auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column-start')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column-end')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row')", "'auto / auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row-start')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row-end')", "'auto'");

    debug("");
    debug("Test getting and setting invalid grid-column and grid-row through JS");
    testColumnRowInvalidJSParsing("4 5", "5 8");
    testColumnRowInvalidJSParsing("4 /", "5 /");
    testColumnRowInvalidJSParsing("5 / none", "8 / foobar");
    testColumnRowInvalidJSParsing("5 5", "8 auto");
    testColumnRowInvalidJSParsing("5 / /", "8 / /");

    // 0 is invalid.
    testColumnRowInvalidJSParsing("0 / 5", "0 / 6");
    testColumnRowInvalidJSParsing("6 / 0", "8 / 0");
    testColumnRowInvalidJSParsing("0", "0");

    testColumnRowInvalidJSParsing("span span / span span", "span span / span span");

    // Negative integer or 0 are invalid.
    testColumnRowInvalidJSParsing("span -1 / -2 span", "-3 span / span -4");
    testColumnRowInvalidJSParsing("0 span / span 0", "span 0 / 0 span");

    // Spans for both initial / end values are not allowed.
    testColumnRowInvalidJSParsing("5 span / span 2", "span 4 / 3 span");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
